Fix up old interoperability tests and move them to their own file.
